move e2e test code to own folder and some common helpers to helpers
